add symbol search input
