+++ app/routes/trade/symbol/symbolsearch/symbollist/symbollist.module.css
@@ -1,0 +1,62 @@
+
+
+.symbolListWrapper {
+  position: absolute;
+  top: 120%;
+  border-radius: 0.2rem;
+  left: 0;
+  z-index: 1000;
+  width: 40rem;
+  background-color: var(--dark3) !important;
+  padding: 0.5rem;
+  border-radius: 0.2rem !important;
+}
+
+.symbolListWrapper div{
+  background-color: transparent !important;
+}
+
+.symbolList{
+  max-height: 16rem;
+  overflow-y: auto;
+  border-radius: 0 !important;
+  scrollbar-width: none;
+}
+
+.symbolListSearch{
+  width: 100%;
+  border-radius: 0.2rem;
+  border: none;
+  position: relative;
+  margin-bottom: 0.5rem;
+}
+
+.symbolListSearchIcon{
+  position: absolute;
+  left: 1rem;
+  top: 50%;
+  transform: translateY(-50%);
+}
+
+.symbolListSearch input{
+  width: 100%;
+  padding: 0.5rem;
+  border-radius: 0.2rem;
+  background-color: var(--dark4);
+  outline: none;
+  border: none;
+  padding-left: 2rem;
+  color: var(--text1);
+}
+
+
+.symbolListItem{
+  cursor: pointer;
+  color: var(--text1);
+}
+
+.symbolListItem:hover{
+  background-color: var(--dark5) !important;
+  border-radius: 0 !important;
+}
+
